Inventory base table view
