+++ routes/web.php
@@ -29,7 +29,7 @@ App\Http\Modules\Account\Routes::all();
 // App\Http\Modules\Reservation\Routes::all();
 App\Http\Modules\Maintenance\Routes::all();
 App\Http\Modules\Ticket\Ticket::routes();
+App\Http\Modules\Inventory\Inventory::routes();
 
-// require_once(base_path('routes/partials/inventory.php'));
 // require_once(base_path('routes/partials/software-inventory.php'));
 // require_once(base_path('routes/partials/workstation.php'));
